I lifted the state from cart details to app component to fix the bug removing from cart
